use SDK to print container logs to stdout
